requirements.txt
attrs==20.3.0
opencv-python==4.5.1.48
numpy==1.19.3
httpx==0.23.0
fastapi==0.78.0
uvicorn==0.17.6
ffmpeg-python==0.2.0
pydantic~=1.9.1
python-multipart==0.0.5
drawline==0.3.3
